twitter_trends-master/requirements_vm.txt
apache_beam==2.33.0
google-cloud-pubsub==1.7.0
requests==2.26.0
